DDAR

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -11,10 +11,10 @@
 int _atoi(char *s)
 {
 	int i = 0;
-	int result = 0;
+	signed long int result = 0;
 	int sign = 1;
 
-	while ((*(s + i) < '0' || *(s + i) > '9') && *(s + i ) != '\0')
+	while ((*(s + i) < '0' || *(s + i) > '9') && *(s + i) != '\0')
 	{
 		if (*(s + i) == '-')
 		{
@@ -22,7 +22,7 @@ int _atoi(char *s)
 		}
 		i++;
 	}
-	while (*(s + i) >= '0' && *(s + i) <= '9' && *(s + i ) != '\0')
+	while (*(s + i) >= '0' && *(s + i) <= '9' && *(s + i) != '\0')
 	{
 		result = result * 10 + (s[i] - '0');
 		i++;
